+++ nexusutils/nyaml2nxdl/nyaml2nxdl_backward_tools.py
@@ -22,12 +22,214 @@
 #
 import sys
 from nexusutils.dataconverter import helpers
+from typing import List
 from . import check_escape_sequence_in_text
 
 
 OPSSIBLE_DIM_ATTRS = ['dim', 'ref', 'optional', 'recommended']
 OPSSIBLE_DIMENSION_ATTRS = ['rank', 'doc']
 DEPTH_SIZE = "  "
+
+
+class Nxdl2yaml():
+    """
+        Parse XML file and print a YML file
+    """
+
+    def __init__(
+            self,
+            symbol_list: List[str],
+            root_level_definition: List[str],
+            root_level_doc='',
+            root_level_symbols=''):
+
+#         self.yaml_dict: dict = {}
+        # updated part of yaml_dict
+#        self.yaml_dict_entrance: dict = self.yaml_dict
+        self.append_flag = True
+        self.found_definition = False
+        self.jump_symbol_child = False
+        self.root_level_doc = root_level_doc
+        self.root_level_symbols = root_level_symbols
+        self.root_level_definition = root_level_definition
+        self.symbol_list = symbol_list
+
+
+
+
+
+
+
+    def handle_symbols(self, depth, node):
+        """Handle symbols field and its childs
+
+        """
+        # pylint: disable=consider-using-f-string
+        self.root_level_symbols = (
+            f"{helpers.remove_namespace_from_tag(node.tag)}: "
+            f"{node.text.strip() if node.text else ''}"
+        )
+        depth += 1
+        for child in list(node):
+            tag = helpers.remove_namespace_from_tag(child.tag)
+            if tag == 'doc':
+                self.symbol_list.append(handle_not_root_level_doc(depth,
+                                                                  text=child.text))
+            elif tag == 'symbol':
+                if 'doc' in child.attrib:
+                    self.symbol_list.append(handle_not_root_level_doc(depth,
+                                                                      tag=child.attrib['name'],
+                                                                      text=child.attrib['doc']))
+                else:
+                    for symbol_doc in list(child):
+                        tag = helpers.remove_namespace_from_tag(symbol_doc.tag)
+                        if tag == 'doc':
+                            self.symbol_list.append(handle_not_root_level_doc(depth,
+                                                                              tag=child.attrib['name'],
+                                                                              text=symbol_doc.text))
+
+    def handle_definition(self, node):
+        """Handle definition group and its attributes
+
+        """
+        # pylint: disable=consider-using-f-string
+        attribs = node.attrib
+        for item in attribs:
+            if 'schemaLocation' not in item \
+                    and 'name' not in item \
+                    and 'extends' not in item \
+                    and 'type' not in item:
+                self.root_level_definition.append(
+                    '{key}: {value}'.format(
+                        key=item,
+                        value=attribs[item] or ''))
+        if 'name' in attribs.keys():
+            self.root_level_definition.append(
+                '{name}:'.format(
+                    name=attribs['name'] or ''))
+            if 'extends' in attribs.keys() and attribs["extends"] != "NXobject":
+                keyword = self.root_level_definition.pop()
+                self.root_level_definition.append(f"{keyword[0:-1]}({attribs['extends']}):")
+
+    def handle_root_level_doc(self, node):
+        """Handle the documentation field found at root level"""
+        # pylint: disable=consider-using-f-string
+
+        for child in list(node):
+            tag = helpers.remove_namespace_from_tag(child.tag)
+            if tag == ('doc'):
+                self.root_level_doc = '{indent}{tag}: | {text}\n'.format(
+                    indent=0 * '  ',
+                    tag=helpers.remove_namespace_from_tag(child.tag),
+                    text='\n' + '  ' + '\n'.join([f"{1 * '  '}{s.lstrip()}"
+                                                  for s in child.text.split('\n')]
+                                                 if child.text else '').strip())
+                node.remove(child)
+
+    def print_root_level_doc(self, file_out):
+        """
+        Print at the root level of YML file \
+        the general documentation field found in XML file
+        """
+        # pylint: disable=consider-using-f-string
+        file_out.write(
+            '{indent}{root_level_doc}'.format(
+                indent=0 * '  ',
+                root_level_doc=self.root_level_doc))
+        self.root_level_doc = ''
+
+    def print_root_level_info(self, depth, file_out):
+        """
+        Print at the root level of YML file \
+        the information stored as definition attributes in the XML file
+        """
+        # pylint: disable=consider-using-f-string
+        if depth >= 0 \
+                and ([s for s in self.root_level_definition if "category: application" in s]\
+                or [s for s in self.root_level_definition if "category: base" in s]):
+            if self.root_level_symbols:
+                file_out.write(
+                    '{indent}{root_level_symbols}\n'.format(
+                        indent=0 * '  ',
+                        root_level_symbols=self.root_level_symbols))
+                for symbol in self.symbol_list:
+                    file_out.write(
+                        '{indent}{symbol}\n'.format(
+                            indent=0 * '  ',
+                            symbol=symbol))
+            if self.root_level_definition:
+                for defs in self.root_level_definition:
+                    file_out.write(
+                        '{indent}{defs}\n'.format(
+                            indent=0 * '  ',
+                            defs=defs))
+            self.found_definition = False
+
+    def recursion_in_xml_tree(self, depth, xml_tree, output_yml, verbose):
+        """
+            Descend lower level in xml tree. If we are in the symbols branch, the recursive
+        behaviour is not triggered as we already handled the symbols' childs.
+        """
+
+        tree = xml_tree['tree']
+        node = xml_tree['node']
+        if self.jump_symbol_child is True:
+            self.jump_symbol_child = False
+        else:
+            for child in list(node):
+                xml_tree_children = {'tree': tree, 'node': child}
+                Nxdl2yaml.xmlparse(self, output_yml, xml_tree_children, depth, verbose)
+
+    def xmlparse(self, output_yml, xml_tree, depth, verbose):
+        """Main of the nxdl2yaml converter.
+It parses XML tree,
+then prints recursively each level of the tree
+
+    """
+        tree = xml_tree['tree']
+        node = xml_tree['node']
+        # TODO remove Nxdl2yaml with self object
+        if verbose:
+            sys.stdout.write(f'Node tag: {helpers.remove_namespace_from_tag(node.tag)}\n')
+            sys.stdout.write(f'Attributes: {node.attrib}\n')
+        with open(output_yml, "a", encoding="utf-8") as file_out:
+            tag = helpers.remove_namespace_from_tag(node.tag)
+            if tag == ('definition'):
+                self.found_definition = True
+                Nxdl2yaml.handle_definition(self, node)
+            if depth == 0 and not self.root_level_doc:
+                Nxdl2yaml.handle_root_level_doc(self, node)
+            if tag == ('doc') and depth != 1:
+                parent = get_node_parent_info(tree, node)[0]
+                doc_parent = helpers.remove_namespace_from_tag(parent.tag)
+                if doc_parent != 'item':
+                    handle_not_root_level_doc(depth, node.text,
+                                              tag=node.tag,
+                                              file_out=file_out)
+            if tag == ('symbols'):
+                Nxdl2yaml.handle_symbols(self, depth, node)
+                self.jump_symbol_child = True
+            # End of root level definition parsing. Print root-level definitions in file
+            # TODO: remove unecessary append_flag
+            if self.root_level_doc \
+                    and self.append_flag is True \
+                    and (depth in (0, 1)):
+                Nxdl2yaml.print_root_level_doc(self, file_out)
+            if self.found_definition is True and self.append_flag is True:
+                Nxdl2yaml.print_root_level_info(self, depth, file_out)
+            # End of print root-level definitions in file
+            if tag in ('field', 'group') and depth != 0:
+                handle_group_or_field(depth, node, file_out)
+            if tag == ('enumeration'):
+                handle_enumeration(depth, node, file_out)
+            if tag == ('attribute'):
+                handle_attributes(depth, node, file_out)
+            if tag == ('dimensions'):
+                handle_dimension(depth, node, file_out)
+        depth += 1
+        # Write nested nodes
+        Nxdl2yaml.recursion_in_xml_tree(self, depth, xml_tree, output_yml, verbose)
+
 
 def handle_not_root_level_doc(depth, text, tag='doc', file_out=None):
     """
@@ -112,6 +314,7 @@ def handle_group_or_field(depth, node, file_out):
             '{indent}unit: {value}\n'.format(
                 indent=(depth + 1) * '  ',
                 value=node.attrib['units'] or ''))
+
 
 # TODO make code radable by moving rank in a rank variable
 def handle_dimension(depth, node, file_out):
